Worker (#10)
* Worker init

* worker service draft

* Worker

* Fix

* Change balance type and add index for expired_at

* Implement new storage methods

* Add serverId to UpdateSubscriptionsBandwidthByKeyID query

* Implement repo

* Worker run

* Fix for review

---------

+++ migrations/001_create_users.up.sql
@@ -1,14 +1,14 @@
 CREATE TABLE IF NOT EXISTS users
 (
     id         BIGSERIAL PRIMARY KEY,
-    first_name TEXT      NOT NULL,
+    first_name TEXT           NOT NULL,
     username   TEXT,
-    balance    INTEGER   NOT NULL DEFAULT 0,
+    balance    NUMERIC(12, 2) NOT NULL DEFAULT 0,
     phone      TEXT,
     partner_id BIGINT,
-    bonus_used BOOLEAN   NOT NULL DEFAULT FALSE,
-    created_at TIMESTAMP NOT NULL DEFAULT NOW(),
-    updated_at TIMESTAMP NOT NULL DEFAULT NOW(),
+    bonus_used BOOLEAN        NOT NULL DEFAULT FALSE,
+    created_at TIMESTAMP      NOT NULL DEFAULT NOW(),
+    updated_at TIMESTAMP      NOT NULL DEFAULT NOW(),
     CONSTRAINT valid_balance CHECK (balance >= 0)
 );
